allsoc: dm: Fix MISRA.ETYPE.CATEGORY.DIFFERENT.2012
The operands of the operator '==' do not have the
same essential type category: one is 'signed' and
the other is 'enum<i>'

Cast SCISERVER_TASK_USER_HI and SCISERVER_TASK_USER_LO
enum values to int32_t to match the essential type of
the task_id field being compared against

Fixes: SITSW-10846

+++ source/drivers/device_manager/sciserver/sciserver_tirtos.c
@@ -346,11 +346,11 @@ static int32_t Sciserver_tirtosInitUserTasks(Sciserver_TirtosCfgPrms_t *sciserve
     {
         TaskP_Params_init (&gSciserverUserTaskParams[i]);
 
-        if (gSciserverTaskList[i].task_id == SCISERVER_TASK_USER_HI)
+        if (gSciserverTaskList[i].task_id == (int32_t)SCISERVER_TASK_USER_HI)
         {
             gSciserverUserTaskParams[i].stack = sciserverCfg->hiTaskStack;
         }
-        else if (gSciserverTaskList[i].task_id == SCISERVER_TASK_USER_LO)
+        else if (gSciserverTaskList[i].task_id == (int32_t)SCISERVER_TASK_USER_LO)
         {
             gSciserverUserTaskParams[i].stack = sciserverCfg->loTaskStack;
         }
